+++ src/components/avatar/avatar.tsx
@@ -1,5 +1,5 @@
 /* eslint-disable no-use-before-define */
-import styled from 'styled-components';
+import styled, { css } from 'styled-components';
 import * as React from 'react';
 
 interface StyleProps {
@@ -11,6 +11,8 @@ interface StyleProps {
    * Fallback color incase image fails to load
    */
   backgroundColor: string;
+
+  stories?: boolean;
 }
 
 export interface AvatarProps {
@@ -26,15 +28,24 @@ const StyledAvatar = styled.img<StyleProps>`
   object-fit: cover;
   border-radius: 100%;
   background-color: ${(props) => props.backgroundColor};
+  ${(props) =>
+    props.stories === true &&
+    css`
+      border: 5px solid #fff;
+      /* offset-x | offset-y | blur-radius | spread-radius | color */
+      box-shadow: 0 0 0 2px orange;
+    `}
 `;
 
 export const Avatar: React.FC<AvatarProps & StyleProps> = ({
   size,
   imageUrl,
   backgroundColor,
+  stories,
 }) => {
   return (
     <StyledAvatar
+      stories={stories}
       size={size}
       src={imageUrl}
       backgroundColor={backgroundColor}
@@ -45,5 +56,6 @@ export const Avatar: React.FC<AvatarProps & StyleProps> = ({
 Avatar.defaultProps = {
   size: 48,
   backgroundColor: 'pink',
-  imageUrl: 'https://picsum.photos/200?a',
+  imageUrl: 'https://source.unsplash.com/200x200/?portrait',
+  stories: false,
 };