+++ src/components/image/image.test.tsx
@@ -1,0 +1,7 @@
+import { Image } from '.';
+
+describe('image', () => {
+  it('is truthy', () => {
+    expect(Image).toBeTruthy();
+  });
+});